+++ sdk/resources/arm-resourcesdeployments/samples-dev/deploymentsCreateOrUpdateAtManagementGroupScopeSample.ts
@@ -1,39 +1,30 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-/**
- * This sample demonstrates how to You can provide the template and parameters directly in the request or link to JSON files.
- *
- * @summary You can provide the template and parameters directly in the request or link to JSON files.
- * x-ms-original-file: specification/resources/resource-manager/Microsoft.Resources/deployments/stable/2025-04-01/examples/PutDeploymentAtManagementGroup.json
- */
-
-import {
-  ScopedDeployment,
-  DeploymentsClient,
-} from "@azure/arm-resourcesdeployments";
+import { ResourcesClient } from "@azure/arm-resourcesdeployments";
 import { DefaultAzureCredential } from "@azure/identity";
-import "dotenv/config";
 
+/**
+ * This sample demonstrates how to you can provide the template and parameters directly in the request or link to JSON files.
+ *
+ * @summary you can provide the template and parameters directly in the request or link to JSON files.
+ * x-ms-original-file: 2025-04-01/PutDeploymentAtManagementGroup.json
+ */
 async function createDeploymentAtManagementGroupScope(): Promise<void> {
-  const groupId = "my-management-group-id";
-  const deploymentName = "my-deployment";
-  const parameters: ScopedDeployment = {
-    location: "eastus",
-    properties: {
-      mode: "Incremental",
-      parameters: {},
-      templateLink: { uri: "https://example.com/exampleTemplate.json" },
-    },
-  };
   const credential = new DefaultAzureCredential();
-  const client = new DeploymentsClient(credential);
-  const result =
-    await client.deployments.beginCreateOrUpdateAtManagementGroupScopeAndWait(
-      groupId,
-      deploymentName,
-      parameters,
-    );
+  const client = new ResourcesClient(credential);
+  const result = await client.deployments.createOrUpdateAtManagementGroupScope(
+    "my-management-group-id",
+    "my-deployment",
+    {
+      location: "eastus",
+      properties: {
+        mode: "Incremental",
+        parameters: {},
+        templateLink: { uri: "https://example.com/exampleTemplate.json" },
+      },
+    },
+  );
   console.log(result);
 }