+++ sdk/resources/arm-resourcesdeployments/CHANGELOG.md
@@ -1,4 +1,52 @@
 # Release History
+
+## 1.0.0-beta.2 (2026-05-15)
+Compared with version 1.0.0-beta.1
+
+### Features Added
+  - Added operation DeploymentsOperations.createOrUpdate
+  - Added operation DeploymentsOperations.createOrUpdateAtManagementGroupScope
+  - Added operation DeploymentsOperations.createOrUpdateAtScope
+  - Added operation DeploymentsOperations.createOrUpdateAtSubscriptionScope
+  - Added operation DeploymentsOperations.createOrUpdateAtTenantScope
+  - Added operation DeploymentsOperations.delete
+  - Added operation DeploymentsOperations.deleteAtManagementGroupScope
+  - Added operation DeploymentsOperations.deleteAtScope
+  - Added operation DeploymentsOperations.deleteAtSubscriptionScope
+  - Added operation DeploymentsOperations.deleteAtTenantScope
+  - Added operation DeploymentsOperations.validate
+  - Added operation DeploymentsOperations.validateAtManagementGroupScope
+  - Added operation DeploymentsOperations.validateAtScope
+  - Added operation DeploymentsOperations.validateAtSubscriptionScope
+  - Added operation DeploymentsOperations.validateAtTenantScope
+  - Added operation DeploymentsOperations.whatIf
+  - Added operation DeploymentsOperations.whatIfAtManagementGroupScope
+  - Added operation DeploymentsOperations.whatIfAtSubscriptionScope
+  - Added operation DeploymentsOperations.whatIfAtTenantScope
+  - Added Class ResourcesClient
+  - Added Interface ExtensionResource
+  - Added Interface PagedAsyncIterableIterator
+  - Added Interface PageSettings
+  - Added Interface Resource
+  - Added Interface ResourcesClientOptionalParams
+  - Added Interface RestorePollerOptions
+  - Added Interface SimplePollerLike
+  - Added Interface SystemData
+  - Added Interface WhatIfOperationProperties
+  - Interface DeploymentExtended has a new optional parameter systemData
+  - Added Type Alias AzureSupportedClouds
+  - Added Type Alias CreatedByType
+  - Added Enum AzureClouds
+  - Added Enum KnownCreatedByType
+  - Added Enum KnownVersions
+
+### Breaking Changes
+  - Deleted Class DeploymentsClient
+  - Removed Interface DeploymentExtendedFilter
+  - Removed Interface DeploymentsClientOptionalParams
+  - Removed Interface ResourceProviderOperationDisplayProperties
+  - Removed Interface SubResource
+
     
 ## 1.0.0-beta.1 (2025-07-03)